requirements.txt
python-dotenv
ipykernel
pandas
numpy
certifi
pymongo[srv]>=4.1,<5.0
scikit-learn
pyaml
mlflow

# -e .